requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
python-dotenv
python-jose
aiofiles
requests
pytest 
jinja
websockets
